+++ src/Operators/Operator.hpp
@@ -7,5 +7,5 @@ using namespace std;
 class Operator {
 public:
     Operator() {};
-    virtual void RenderUpdate(float deltaTime) {throw logic_error("Operator not implemented");};
+    virtual void RenderUpdate(double deltaTime) {throw logic_error("Operator not implemented");};
 };
